requirements.txt
dnspython
argparse
shodan
requests
tldextract
tqdm